+++ lib/sprout/src/main/kotlin/org/partiql/sprout/generator/target/kotlin/KotlinOptions.kt
@@ -9,4 +9,5 @@ class KotlinOptions(
     val packageRoot: String,
     val poems: List<String>,
     val optIns: List<String> = emptyList(),
+    val restrictInterfaceImpl: Boolean = false,
 )