+++ CHANGELOG.md
@@ -33,6 +33,8 @@ Thank you to all who have contributed!
 ### Changed
 - Change `StaticType.AnyOfType`'s `.toString` to not perform `.flatten()`
 - Change modeling of `COALESCE` and `NULLIF` to dedicated nodes in logical plan
+- partiql-plan: Adds an `OptIn` annotation `DoNotImplementInterface` for the `PlanVisitor` interface
+  - `PlanVisitor` should not be implemented directly in favor of extending `PlanBaseVisitor`
 
 ### Deprecated
 - The current SqlBlock, SqlDialect, and SqlLayout are marked as deprecated and will be slightly changed in the next release.